# Conflicts:
#	admin/section/class-convertkit-settings-general.php
#	composer.json
#	views/backend/setup-wizard/convertkit-setup/content-1.php
#	wp-convertkit.php

+++ admin/class-multi-value-field-table.php
@@ -182,7 +182,7 @@ class Multi_Value_Field_Table extends WP_List_Table {
 			array(
 				'total_items' => $total_items,
 				'per_page'    => $per_page,
-				'total_pages' => ceil( $total_items / $per_page ),
+				'total_pages' => (int) ceil( $total_items / $per_page ),
 			)
 		);
